🔧 refactor ConversationService to enhance modularity and response handling
- Introduce **RagContextService** and **ResponseInsertionService** to separate concerns, improving code organization and maintainability.
- Update **ConversationService** to utilize the new services for context retrieval and response insertion, streamlining the AI interaction process.
- Remove redundant methods from **ConversationService** to adhere to the single responsibility principle, enhancing clarity and reducing complexity.
- Ensure consistent error handling and improve the overall user experience by leveraging the new modular structure.
